refactor(sftp): move SideSftp component to FileTrans.tsx and rename to FileTrans
